Turret State Machine and Initial Shooting
I have added a State Machine to the Turret, featuring the states of:
- Idle: Do nothing (Will eventually rotate between two points)
- Alert: Look towards the player
- Shooting: Fire at the player
- Deactivated: Wait until should reactivate)

I have also added the start of the turret shooting, where it currently fires a three round burst at the player, never missing unless they are behind cover. I will soon change this to use the same kind of code as the weapons.