Weapon Work
Note: We are doing it this way so that no Scriptable Objects have changing values, allowing the same Scriptable Object values to be used multiple times throughout a scene.

We have changed some of the values to allow for Scriptable Objects to be used in place of data on the Gun script, allowing us to reuse the values. additionally, we have added recoil over time, and a more advanced recoil system that uses a Texture2D.

Removed Features:
- Changing Fire Types. We have momentarily removed changing firing types for ease of when we were testing.

Current Gun Features:
- Damage (Now SO)
- Shooting (Now SO)
- Fire Rate
- Firing Types
- Bullet Trails
- Ammo & Reloading
- Bullet Impacts
- Alternate Fire
- Bullet Spread