Player Health UI
New Features:
- Progress Bar. I have added a progress bar to the game, which can be used for things such as interaction progress, reload progress, or the player's health bar.
- Player Health Bar. I have added a health bar to the player, allowing you to visualize how much damage you have taken.

Note: We can improve on the progress bar's drain over time, as I don't think it works when we increase the value, and it currently feels too slow. We can also add options for where there are two bars: One that instantly transitions and another that transitions over time, then layer them on top of one another.