Factions and ReadOnly Attribute
I have added the ability to assign entities to Factions, allowing for enemies to differentiate friendly units from enemy ones.

I have also added a ReadOnly attribute, which allows us to display data in the inspector that we wish to not be editable.
I have also added the ability to toggle the debug Gizmos and Logs of the Turret enemy, though not all are hooked up to these two booleans yet.