Settings Menu Scripting
I have done a lot of work and we now have a working settings menu!

This allows us to edit the game's audio volume (Currently not set up with the AudioSources), video settings, crosshair settings and control settings (For both MnK & Gamepad).

We will next need to do some assorted fixes for stuff like AudioMixers and navigating the UI with a controller, but after that we can get started on rebinding keys!

Notes:
- Manager. We have removed the pause functionality from the Manager script.
- PlayerManager. We have added a PlayerManager, now making it responsible for holding values such as the player's sensitivity & crosshair colour.
- Crosshair Colour. We are currently using an Enumerator for determining the crosshair's colour. This is fine, but it may lead to issues if we want to expand the colours available. A static struct containing a dictionary of Colours with string identifiers could work.