Enemy Targeting, Collateral Damage, & Wander Start
I have added enemy senses and targeting to the Standard Enemies, along with the ability for attacks to have collateral damage toggleable by a boolean (Able to damage entities in the same faction).

Furthermore, I have started work on the Standard Enemy's wander state, which will be the state it will enter when it doesn't have a target.
An alternative idea to this could be having each enemy have a reference to the player's transform, assigned when they are spawned, and if they don't have a current target, we pass that instead. This would make all enemies move to the player unless they already have a target, such as a turret.