AttackingState Removal & Hurtbox Attack Fix
I have removed the AttackingState, instead integrating its abilities into the AttackingTarget class, as it was the only one that inherited from it.

Additionally, I have made Hurtboxes ensure that the collider they collide with is NOT a trigger collider before triggering the OnColliderEnter event, which has removed the bug where the enemy would hit the player's interaction sphere collider as well, hitting them twice and dealing double damage.

Finally, I have also increased the range and intensity of the flashlight